msm: kgsl: Reduce latency while processing ioctls
The GPU's ioctls sleep a lot due to memory allocation, during which time
the CPU can enter a deep idle state and take a while to finish executing
the ioctl afterwards. We can alleviate this by optimistically assuming
that the task running the ioctl won't migrate CPUs while it runs and
restricting the current CPU to shallow idle states.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Fiqri Ardyansyah <fiqri15072019@gmail.com>
Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>
Signed-off-by: neophyte <neophyteplayground@gmail.com>